simplify building of executables (#19)
* use choco install aqt
there isn't a qt6 on choco (qt5-default, but not qt6)

* use qt action to install for windows

* get rid of trying to find and include Qt Assistant.
If they have it on their system, they get the help.  If they don't... at least they get the app.
Finding and installing Qt Assistant is failing on all platforms, and isn't absolutely critical.

* use bash when invoking poetry
use powershell commands for windows if not invoking poetry

* fixed creation of log path

* fix up log_path type and usage
constrain executable build to windows

* switch to one file executable

* fix creation of distribution package:  simplify to just zipping up the executable

* remove the .exe file after compressing

* try out not zipping, it seems like the upload of artefacts does that on it's own